// src/lib/framework/Voyado/types/IVoyado.ts
import {
  type VoyadoPromotion,
  type VoyadoTransactions,
  type VoyadoVoucherItem,
  type VoyadoVouchers,
} from '@/src/lib/types/voyado';

export namespace IVoyado {
  export type ContactByEmail = string;

  export interface CreateContact extends Record<string, unknown> {
    email: string;
    firstName?: string | null;
    lastName?: string | null;
    isRegistrationCompleted?: boolean;
    phone?: string;
    contactType?: Contact['meta']['contactType'];
  }

  export interface CreateContactBody extends CreateContact {
    countryCode: string;
    language: string | undefined;

    consents: Array<{
      id: string;
      value: boolean;
      source: 'ecom';
    }>;
  }

  export interface Contact {
    id: string;

    attributes: Record<string, unknown> & {
      email: string;
      firstName: string | null;
      lastName: string | null;

      language: string | null;

      memberNumber: string;
      mobilePhone?: string;
      bonusBasedLevel: string;
      bonusBasedLevelChanged: string;
      bonusBasedLevelExpires: string | null;
      bonusBasedLevelLeftForUpgrade: number;
      memberLevelsBonusPoints: number;
      bonusPoints: number;

      acceptsEvents: boolean;
      acceptsNewsletters: boolean;
      acceptsCampaigns: boolean;
      acceptsMemberCampaigns: boolean;

      /**
       * @deprecated Seems like customer didn't fill out this correctly before
       */
      isRegistrationCompleted: boolean;
    };

    meta: Record<string, unknown> & {
      /**
       * `Member` - Completely registered member.
       *
       * `Contact` - Only the newsletter subscriber.
       *
       * `transactContact` - Automatically generated by completing an order in the system.
       *
       * @see [Asana - Voyado contactTypes](https://app.asana.com/0/1208043166295182/1208465096592203/f)
       */
      contactType: 'Member' | 'Contact' | 'transactContact';

      createdOn: string;
      modifiedOn: string;
      sourceType: string;
      approved: true;
    };

    preferences: {
      AcceptsEmails: boolean;
      AcceptsPostals: boolean;
      AcceptsSms: boolean;
    };

    consents: unknown[];
  }

  export interface ContactOverview extends Omit<Contact, 'preferences'> {
    attributes: Contact['attributes'] & {
      street: string | null;
      zipCode: string | null;
      city: string | null;
      countryCode: string | null;
      careOf: string | null;
      currentStore: {
        id: string;
        name: string;
        externalId: string;
      } | null;
      rfm: unknown | null;
      discoveryKey: string | null;
      lastOrderDate: string | null;
      lang: string | null;
      secrecyMarked: boolean;
    };
    meta: Contact['meta'] & {
      recruitedByStore: {
        id: string;
        name: string;
        externalId: string;
      } | null;
    };
    preferences: {
      acceptsEmail: boolean;
      acceptsPostal: boolean;
      acceptsSms: boolean;
    };
    consents: {
      id: string;
      value: boolean;
      date: string;
      comment: string;
      source: string;
    }[];
    warnings: unknown[];
    audiences: unknown[];
    bonusChecks: VoyadoVoucherItem[];
    promotions: Promotions;
  }

  export type Promotions = VoyadoPromotion[];

  export type Transactions = VoyadoTransactions;

  export type Vouchers = VoyadoVouchers;
  export type Subscription = {
    ContactId: string;
    Sku: string;
    Locale: string;
    ExternalId?: string;
  };

  export interface Store {
    name: string;
    countryCode: string;
    externalId: string;
  }

  export interface Consent {
    id: string;
    name: string;
  }

  export interface MultipleMatchesError {
    errorCode: 'MultipleMatches';
    messageDetails: {
      multipleMatchesFound: string[];
    };
    message: string;
    requestId: string;
  }
}